refactor: rename encoding_processor -> save_actor
