+++ src/services/statefuls/cache/manager.rs
@@ -6,8 +6,8 @@ use crate::services::cluster::replications::replication::ReplicationInfo;
 use crate::services::query_io::QueryIO;
 use crate::services::statefuls::cache::ttl::actor::TtlActor;
 use crate::services::statefuls::cache::ttl::manager::TtlSchedulerInbox;
-use crate::services::statefuls::snapshot::endec::encoder::encoding_processor::SaveActor;
-use crate::services::statefuls::snapshot::endec::encoder::encoding_processor::SaveTarget;
+use crate::services::statefuls::snapshot::save_actor::SaveActor;
+use crate::services::statefuls::snapshot::save_actor::SaveTarget;
 use crate::services::statefuls::snapshot::DumpFile;
 use anyhow::Result;
 use std::time::SystemTime;
