+++ src/services/statefuls/snapshot/endec/encoder/mod.rs
@@ -1,2 +1,1 @@
 pub mod byte_encoder;
-pub(crate) mod encoding_processor;
