+++ src/services/statefuls/snapshot/mod.rs
@@ -1,6 +1,8 @@
 pub mod dump_loader;
 pub mod encoding_command;
 pub mod endec;
+pub(crate) mod save_actor;
+
 use super::cache::CacheEntry;
 
 #[derive(Debug)]
